added some styling to the individual game page, only the game information
