+++ react-app/src/components/Review.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
-import { getAllReviewsPerGame, eraseReview } from "../store/review";
+import { getAllReviewsPerGame } from "../store/review";
 import ReviewEditForm from "./ReviewEditForm";
 import ReviewDelete from "./ReviewDelete";
