+++ react-app/src/index.css
@@ -1,1 +1,5 @@
 /* TODO Add site wide styles */
+* {
+  background-color: rgb(47, 41, 49);
+  color: white;
+}
